refactor/meta_voting: move MetaEvent to separate file
